Extract cached menu data from NavLinks and footer
Move "use cache" GraphQL fetches into lib/menus and lib/channels data
modules; layout components become thin wrappers plus presentational views.
