+++ src/lib/channels/get-channels-data.ts
@@ -1,0 +1,20 @@
+import { ChannelsListDocument, type ChannelsListQuery } from "@/gql/graphql";
+import { executePublicGraphQL } from "@/lib/graphql";
+import { CACHE_PROFILES, applyCacheProfile } from "@/lib/cache-manifest";
+
+export async function getCachedChannelsList(): Promise<ChannelsListQuery | null> {
+	"use cache";
+	applyCacheProfile(CACHE_PROFILES.channels);
+
+	if (!process.env.SALEOR_APP_TOKEN) {
+		return null;
+	}
+
+	const result = await executePublicGraphQL(ChannelsListDocument, {
+		headers: {
+			Authorization: `Bearer ${process.env.SALEOR_APP_TOKEN}`,
+		},
+	});
+
+	return result.ok ? result.data : null;
+}
