+++ src/ui/components/nav/components/nav-links.tsx
@@ -1,58 +1,7 @@
-import Link from "next/link";
-import { NavLink } from "./nav-link";
-import { executePublicGraphQL } from "@/lib/graphql";
-import { MenuGetBySlugDocument } from "@/gql/graphql";
-import { CACHE_PROFILES, applyCacheProfile, NAVBAR_MENU_SLUG } from "@/lib/cache-manifest";
+import { getNavbarMenuItems } from "@/lib/menus/get-menu-data";
+import { NavLinksView } from "./nav-links-view";
 
-export const NavLinks = async ({ channel }: { channel: string }) => {
-	"use cache";
-	applyCacheProfile(CACHE_PROFILES.navigation, { channel });
-
-	const result = await executePublicGraphQL(MenuGetBySlugDocument, {
-		variables: { slug: NAVBAR_MENU_SLUG, channel },
-	});
-
-	if (!result.ok) {
-		// During build, if the API is unreachable, render minimal nav.
-		// The page will re-fetch when a user visits.
-		console.warn(`[NavLinks] Failed to fetch navigation for ${channel}:`, result.error.message);
-		return <NavLink href="/products">All</NavLink>;
-	}
-
-	return (
-		<>
-			<NavLink href="/products">All</NavLink>
-			{result.data.menu?.items?.map((item) => {
-				if (item.category) {
-					return (
-						<NavLink key={item.id} href={`/categories/${item.category.slug}`}>
-							{item.category.name}
-						</NavLink>
-					);
-				}
-				if (item.collection) {
-					return (
-						<NavLink key={item.id} href={`/collections/${item.collection.slug}`}>
-							{item.collection.name}
-						</NavLink>
-					);
-				}
-				if (item.page) {
-					return (
-						<NavLink key={item.id} href={`/pages/${item.page.slug}`}>
-							{item.page.title}
-						</NavLink>
-					);
-				}
-				if (item.url) {
-					return (
-						<Link key={item.id} href={item.url} prefetch={false}>
-							{item.name}
-						</Link>
-					);
-				}
-				return null;
-			})}
-		</>
-	);
-};
+export async function NavLinks({ channel }: { channel: string }) {
+	const items = await getNavbarMenuItems(channel);
+	return <NavLinksView items={items ?? []} />;
+}
